Store credits as whole numbers
Credits were a float, but nothing ever produces a fraction of one. The
listing bonus is 100, a loan costs the daily rate times a whole number
of days. All the float did was risk a balance ending up at 99.99999 and
print "330.0" everywhere.

Changed to int. The casts that were being used to tidy up the messages
are gone too.

+++ backend/src/main/java/com/borrowbox/model/LendingService.java
@@ -29,7 +29,7 @@ public class LendingService {
 
     if (borrower.getCredits() < contract.getCost()) {
       throw new LendingNotAllowedException(
-          "The borrower has " + (int) borrower.getCredits()
+          "The borrower has " + borrower.getCredits()
               + " credits but this loan costs " + contract.getCost() + ".");
     }
